requirements.txt
streamlit
openai
pandas
chromadb
python-dotenv
pyperclip
